+++ MyFinance/templates/voucher/voucher_entry.html
@@ -9,12 +9,12 @@ Voucher Entry
 {% block content %}
 <div class="container">
 
-<!-- Button trigger modal -->
+<!-- Button trigger modal 
 <button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#exampleModal">
     Launch demo modal
   </button>
-  
-  <!-- Modal -->
+-->  
+  <!-- Modal 
   <div class="modal fade" id="exampleModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
     <div class="modal-dialog">
       <div class="modal-content">
@@ -32,7 +32,8 @@ Voucher Entry
       </div>
     </div>
   </div>
-
+  -->
+  
     <form action="/voucher/voucher_result" method="POST">
         <div class="row gx-5">
             <div class="col-auto form-floating">
