+++ docs/source/conf.py
@@ -10,9 +10,9 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
-# import os
-# import sys
-# sys.path.insert(0, os.path.abspath('.'))
+import os
+import sys
+sys.path.insert(0, '/Users/edmundlskoviak/Documents/repos/finance_cmd')
 
 
 # -- Project information -----------------------------------------------------
@@ -31,6 +31,9 @@ release = '0.9'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
+    'sphinx.ext.duration',
+    'sphinx.ext.doctest',
+    'sphinx.ext.autodoc',    
 ]
 
 # Add any paths that contain templates here, relative to this directory.
