+++ docs/source/sections/MyFinance/index.rst
@@ -1,0 +1,20 @@
+=======
+finance
+=======
+
+This is the main application package, containing the Applicaiton Factory ``create_app()`` function of Flask.  This is also the 
+function used as the object for the WSGI app needed in ``wsgi.py``.
+
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+.. autofunction:: MyFinance.finance.create_app
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
